Spy-mobile/requirements.txt
kivy
pyjnius
requests
